+++ duat-core/src/mode/remap.rs
@@ -475,7 +475,7 @@ pub enum Gives {
     Mode(Box<dyn Fn() + Send>),
 }
 
-fn remove_alias_and<U: Ui>(f: impl FnOnce(&mut Text, usize)) {
+fn remove_alias_and<U: Ui>(f: impl FnOnce(&mut Text, u32)) {
     let widget = context::cur_widget::<U>().unwrap();
     widget.mutate_data(|widget, _, cursors| {
         let cursors = cursors.read();
